refactor(exception): create exception package for exception handling
