+++ src/main/java/woowacourse/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.exception;
 
 public class AuthenticationException extends RuntimeException {
