+++ src/main/java/woowacourse/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
 public class NotFoundException extends RuntimeException {
