+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -20,7 +20,7 @@ import woowacourse.shoppingcart.dto.request.UpdateMeRequest;
 import woowacourse.shoppingcart.dto.request.UpdatePasswordRequest;
 import woowacourse.shoppingcart.dto.response.GetMeResponse;
 import woowacourse.shoppingcart.dto.response.UniqueUsernameResponse;
-import woowacourse.shoppingcart.exception.NotFoundException;
+import woowacourse.exception.NotFoundException;
 import woowacourse.util.DatabaseFixture;
 
 @SuppressWarnings("NonAsciiCharacters")
